refactor: move `CCardDialog` to `dialogs`
